roles/client/files/requirements.txt
ansible==2.1.1.0
cffi==1.7.0
cryptography==1.4
Django==1.9.6
django-auth-ldap==1.2.8
docutils==0.12
enum34==1.1.6
extras==1.0.0
gear==0.5.9
idna==2.1
ipaddress==1.0.16
jenkins-job-builder==1.6.0
Jinja2==2.8
lockfile==0.12.2
lxml==3.6.0
MarkupSafe==0.23
multi-key-dict==2.0.3
paramiko==2.0.2
pbr==1.10.0
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
python-daemon==2.1.1
python-jenkins==0.4.12
python-ldap==2.4.25
pytz==2016.4
PyYAML==3.11
requests==2.10.0
simplejson==3.8.2
six==1.10.0
